+++ packages/agents-sdk/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @inkeep/agents-sdk
 
+## 0.47.5
+
+### Patch Changes
+
+- @inkeep/agents-core@0.47.5
+
+## 0.47.4
+
+### Patch Changes
+
+- Updated dependencies [83346fc]
+- Updated dependencies [5f3f5ea]
+  - @inkeep/agents-core@0.47.4
+
+## 0.47.3
+
+### Patch Changes
+
+- Updated dependencies [756a560]
+- Updated dependencies [045c405]
+  - @inkeep/agents-core@0.47.3
+
 ## 0.47.2
 
 ### Patch Changes